Add line filter function
Reorganize the code a bit, as it shares most of the code with the line
filter for brightfield reconstruction: move many functions to the utils
module. Also moves the bf reconstruction code to
`libertem_holo.base.reconstr`.
